refactor: create drinks module
