+++ app/modules/drinks/implementation/types.ts
@@ -1,0 +1,9 @@
+export type EnhancedDrink = {
+  title: string;
+  slug: string;
+  image: { url: string; blurDataUrl: string };
+  ingredients: string[];
+  calories: number;
+  notes: string | null;
+  tags: string[];
+};
