+++ app/modules/drinks/index.ts
@@ -1,0 +1,40 @@
+// Types
+export type { EnhancedDrink } from "./implementation/types";
+
+// Queries
+export {
+  getAllDrinks,
+  getPublishedDrinks,
+  getDrinkBySlug,
+  getDrinksByTag,
+  getAllTags,
+  createDrink,
+  updateDrink,
+  deleteDrink,
+} from "./implementation/queries.server";
+
+// Validation
+export { drinkFormSchema } from "./implementation/validation";
+
+// View helpers
+export { withPlaceholderImages } from "./implementation/placeholder-images.server";
+export { parseImageUpload } from "./implementation/parse-image-upload.server";
+export { markdownToHtml } from "./implementation/markdown.server";
+
+// Image management (used by route actions until mutations consolidation)
+export { uploadImageOrPlaceholder, deleteImage } from "./implementation/imagekit.server";
+
+// Cache
+export { purgeDrinkCache } from "./implementation/fastly.server";
+export { getSurrogateKeyForTag } from "./implementation/tags";
+
+// UI
+export { DrinkDetails } from "./ui/drink-details";
+export { DrinkList } from "./ui/drink-list";
+export { DrinkSummary } from "./ui/drink-summary";
+export { DrinkForm } from "./ui/drink-form";
+export { type ImageCropHandle, ImageCrop } from "./ui/image-crop";
+export { Glass } from "./ui/glass";
+export { Tag } from "./ui/tag";
+export { TagLink } from "./ui/tag-link";
+export { useSortableData } from "./ui/use-sortable-data";
