+++ app/routes/_app.search/route.tsx
@@ -2,8 +2,7 @@ import { data, useSearchParams, useNavigation } from "react-router";
 import { cacheHeader } from "pretty-cache-header";
 import { defaultPageDescription, defaultPageTitle } from "#/app/core/config";
 import { getEnvVars } from "#/app/utils/env.server";
-import { withPlaceholderImages } from "#/app/utils/placeholder-images.server";
-import { DrinkList } from "#/app/drinks/drink-list";
+import { withPlaceholderImages, DrinkList } from "#/app/modules/drinks";
 import type { Drink } from "#/app/db/schema";
 import type { AppRouteHandle } from "#/app/types";
 import { searchDrinks } from "#/app/search/minisearch.server";
