+++ app/routes/admin.drinks.$slug.edit.tsx
@@ -1,13 +1,17 @@
 import { redirect, href, data } from "react-router";
 import { invariantResponse } from "@epic-web/invariant";
-import { getDrinkBySlug, updateDrink } from "#/app/models/drink.server";
-import { uploadImageOrPlaceholder, deleteImage } from "#/app/utils/imagekit.server";
-import { DrinkForm } from "#/app/admin/drink-form";
-import { parseImageUpload } from "#/app/utils/parse-image-upload.server";
+import {
+  getDrinkBySlug,
+  updateDrink,
+  uploadImageOrPlaceholder,
+  deleteImage,
+  DrinkForm,
+  parseImageUpload,
+  purgeDrinkCache,
+  drinkFormSchema,
+} from "#/app/modules/drinks";
 import { purgeSearchCache } from "#/app/search/cache.server";
-import { purgeDrinkCache } from "#/app/utils/fastly.server";
 import { getSession, commitSession } from "#/app/modules/auth";
-import { drinkFormSchema } from "#/app/validation/drink";
 import type { Route } from "./+types/admin.drinks.$slug.edit";
 
 export async function loader({ params }: Route.LoaderArgs) {
